hp9845: hp9845b_state class moved into a separated file (hp9845.h)
